chore: move stories next to components
