+++ frontend/src/components/Experiment/Header/Header.stories.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom";
 
-import Header from "../components/Experiment/Header/Header";
+import Header from "./Header";
 
 export default {
     title: "Header/Header",
