+++ frontend/src/components/FloatingActionButton/FloatingActionButton.stories.tsx
